update shacl validation scripts
